+++ assets/javascripts/kitten/components/images/image-cropper/index.js
@@ -1,13 +1,14 @@
 import React, { useEffect, useState, useRef } from 'react'
+import classNames from 'classnames'
 import Cropper from 'react-cropper'
 import getOr from 'lodash/fp/getOr'
-import { Marger } from '../../../components/layout/marger'
-import { Grid, GridCol } from '../../../components/grid/grid'
 import { Label } from '../../../components/form/label'
 import { Paragraph } from '../../../components/typography/paragraph'
 import { BasicUploader } from '../../../components/uploaders/basic-uploader'
 import { RangeSlider } from '../../../components/form/range-slider'
 import { domElementHelper } from '../../../helpers/dom/element-helper'
+import { StyledCropper } from './styles'
+import { useFlexGapCheck } from '../../../helpers/dom/use-flex-gap-check'
 
 export const ImageCropper = ({
   imageSrc,
@@ -25,6 +26,7 @@ export const ImageCropper = ({
   description,
   cropperInfo,
   sliderTitle,
+  className,
 }) => {
   const cropperContainerRef = useRef(null)
   const cropperRef = useRef(null)
@@ -40,6 +42,7 @@ export const ImageCropper = ({
   const [initialSliderValue, setInitialSliderValue] = useState(0)
   const [uploadedFile, setUploadedFile] = useState(null)
   const [resultData, setResultData] = useState(null)
+  const canUseGap = useFlexGapCheck()
 
   useEffect(() => {
     if (cropperInstance && cropperInstance.imageData.naturalWidth) {
@@ -92,110 +95,123 @@ export const ImageCropper = ({
   }, [resultData, fileNameState, uploadedFile])
   const dragMode = disabled || !isCropEnabled ? 'none' : 'move'
   return (
-    <>
-      <Marger bottom="1.5">
-        <Label size="tiny" htmlFor={name}>
-          {label}
-        </Label>
-      </Marger>
+    <StyledCropper
+      className={classNames('k-UploadAndCropper', className, {
+        'k-UploadAndCropper--noGap': !canUseGap,
+      })}
+    >
+      <Label
+        size="tiny"
+        htmlFor={name}
+        className="k-u-margin-bottom-singleHalf"
+      >
+        {label}
+      </Label>
 
-      <Marger top="1.5" bottom="1">
-        <BasicUploader
-          id={name}
-          fileName={fileNameState}
-          buttonText={buttonLabel}
-          disabled={disabled}
-          errorText={errorText}
-          status={status}
-          fileInputProps={{ accept: acceptedFiles }}
-          onUpload={e => {
-            try {
-              const file = e.currentTarget.files[0]
-              setUploadedFile(file)
-              if (file.size > maxSize) {
-                setStatus('error')
-                setErrorText(uploaderErrorLabel)
-              } else {
-                const reader = new FileReader()
-                reader.onload = event => {
-                  setImageSrc(event.target.result)
-                  setFileName(file.name)
-                }
-                reader.readAsDataURL(file)
-              }
-            } catch (e) {
+      <BasicUploader
+        id={name}
+        className="k-u-margin-top-singleHalf k-u-margin-bottom-single"
+        fileName={fileNameState}
+        buttonText={buttonLabel}
+        disabled={disabled}
+        errorText={errorText}
+        status={status}
+        fileInputProps={{ accept: acceptedFiles }}
+        onUpload={e => {
+          try {
+            const file = e.currentTarget.files[0]
+            setUploadedFile(file)
+            if (file.size > maxSize) {
               setStatus('error')
               setErrorText(uploaderErrorLabel)
+            } else {
+              const reader = new FileReader()
+              reader.onload = event => {
+                setImageSrc(event.target.result)
+                setFileName(file.name)
+              }
+              reader.readAsDataURL(file)
             }
-          }}
-          onCancel={() => {
-            setImageSrc(imageSrc)
-            setFileName(fileName)
-            setErrorText('')
-            setUploadedFile(null)
+          } catch (e) {
+            setStatus('error')
+            setErrorText(uploaderErrorLabel)
+          }
+        }}
+        onCancel={() => {
+          setImageSrc(imageSrc)
+          setFileName(fileName)
+          setErrorText('')
+          setUploadedFile(null)
 
-            onChange({
-              value: null,
-              name: null,
-              file: null,
-            })
-          }}
-        />
-      </Marger>
+          onChange({
+            value: null,
+            name: null,
+            file: null,
+          })
+        }}
+      />
 
-      <Marger top="1">
-        <Paragraph modifier="quaternary" margin={false}>
-          {description}
-        </Paragraph>
-      </Marger>
+      <Paragraph
+        modifier="quaternary"
+        margin={false}
+        className="k-u-margin-top-single"
+      >
+        {description}
+      </Paragraph>
 
-      {imageSrcState && (
-        <Grid>
-          <GridCol col="12" col-m="6">
-            <Marger top="2">
-              <div ref={cropperContainerRef}>
-                {cropperWidth && cropperHeight && (
-                  <Cropper
-                    onInitialized={instance => {
-                      setCropperInstance(instance)
-                    }}
-                    ref={cropperRef}
-                    className="k-Cropper"
-                    src={imageSrcState}
-                    style={styles}
-                    initialAspectRatio={aspectRatio}
-                    viewMode={3}
-                    guides={false}
-                    modal={false}
-                    autoCropArea={1}
-                    cropBoxMovable={false}
-                    cropBoxResizable={false}
-                    toggleDragModeOnDblclick={false}
-                    zoomOnTouch={false}
-                    zoomOnWheel={false}
-                    dragMode={dragMode}
-                    crop={result => {
-                      setResultData(result)
-                    }}
-                  />
-                )}
-              </div>
-            </Marger>
-          </GridCol>
+      <div
+        className="k-Cropper__wrapper k-u-margin-top-double"
+        aria-live="polite"
+      >
+        {imageSrcState && (
+          <>
+            <div
+              ref={cropperContainerRef}
+              className="k-Cropper__wrapper__cropper"
+            >
+              {cropperWidth && cropperHeight && (
+                <Cropper
+                  onInitialized={instance => {
+                    setCropperInstance(instance)
+                  }}
+                  ref={cropperRef}
+                  className="k-Cropper"
+                  src={imageSrcState}
+                  style={styles}
+                  initialAspectRatio={aspectRatio}
+                  viewMode={3}
+                  guides={false}
+                  modal={false}
+                  autoCropArea={1}
+                  cropBoxMovable={false}
+                  cropBoxResizable={false}
+                  toggleDragModeOnDblclick={false}
+                  zoomOnTouch={false}
+                  zoomOnWheel={false}
+                  dragMode={dragMode}
+                  crop={result => {
+                    setResultData(result)
+                  }}
+                />
+              )}
+            </div>
 
-          {isCropEnabled && !disabled && (
-            <GridCol col="12" col-m="6">
-              <Marger top="2" bottom="1.5">
-                <Paragraph modifier="quaternary" margin={false}>
+            {isCropEnabled && !disabled && (
+              <div className="k-Cropper__wrapper__slider">
+                <Paragraph
+                  modifier="quaternary"
+                  margin={false}
+                  className="k-u-margin-bottom-singleHalf"
+                >
                   {cropperInfo}
                 </Paragraph>
-              </Marger>
-              <Marger top="1.5" bottom="1">
-                <Label size="micro" htmlFor="zoomSlider">
+                <Label
+                  size="micro"
+                  htmlFor="zoomSlider"
+                  className="k-u-margin-top-singleHalf k-u-margin-bottom-single"
+                >
                   {sliderTitle}
                 </Label>
-              </Marger>
-              <Marger top="1">
                 <RangeSlider
                   id="zoomSlider"
                   name="zoom"
@@ -207,13 +223,14 @@ export const ImageCropper = ({
                     const value = event.target.value
                     cropperInstance && cropperInstance.zoomTo(value / 100)
                   }}
+                  className="k-u-margin-top-single"
                 />
-              </Marger>
-            </GridCol>
-          )}
-        </Grid>
-      )}
-    </>
+              </div>
+            )}
+          </>
+        )}
+      </div>
+    </StyledCropper>
   )
 }