+++ src/esm/index.js
@@ -8,6 +8,9 @@ export { Accordeon } from './components/accordeon'; // Atoms
 
 export { Badge } from './components/atoms/badge';
 export { IconBadge } from './components/atoms/icon-badge';
+export { Loader } from './components/atoms/loader';
+export { LoaderWithParagraph } from './components/atoms/loader-with-paragraph';
+export { StepperIcon, StepperIconDefault, StepperIconInProgress, StepperIconValidated } from './components/atoms/stepper-icon';
 export { Tag } from './components/atoms/tag'; // Avatar
 
 export { AvatarWithTextAndBadge } from './components/avatar/avatar-with-text-and-badge'; // Box
@@ -96,7 +99,7 @@ export { Grid, GridCol } from './components/grid/grid';
 export { Row } from './components/grid/row';
 export { SideGrid } from './components/grid/side-grid';
 export { StickyContainer } from './components/grid/sticky-container';
-export { AdaptableGrid } from './components/grid/adaptable-grid'; // Heroes
+export { AdaptableGrid, AdaptableGridCol } from './components/grid/adaptable-grid'; // Heroes
 
 export { Hero } from './components/heroes/hero'; // Images
 
@@ -121,10 +124,7 @@ export { IconList } from './components/lists/icon-list';
 export { BulletList } from './components/lists/bullet-list';
 export { List } from './components/lists/list';
 export { CheckList } from './components/lists/check-list';
-export { Timeline } from './components/lists/timeline'; // Loaders
-
-export { LoaderWithParagraph } from './components/loaders/loader-with-paragraph';
-export { Loader } from './components/loaders/loader'; // Menus
+export { Timeline } from './components/lists/timeline'; // Menus
 
 export { DashboardMenu } from './components/menus/dashboard-menu';
 export { HeaderMenu } from './components/menus/header-menu'; // Meters
@@ -153,7 +153,6 @@ export { SearchInput } from './components/search/search-input'; // Status with B
 export { StatusWithBullet } from './components/status-with-bullet'; // Steppers
 
 export { VerticalStepper } from './components/steppers/vertical-stepper';
-export { StepperIconDefault, StepperIconInProgress, StepperIconValidated } from './components/steppers/stepper-icon';
 export { Stepper } from './components/steppers/stepper'; // Tables
 
 export { DoubleEntryTable } from './components/tables/double-entry-table';
@@ -162,11 +161,7 @@ export { SingleEntryTable } from './components/tables/single-entry-table'; // Te
 
 export { TextCopy } from './components/text-copy'; // Tooltips
 
-export { QuestionMarkWithTooltip } from './components/tooltips/question-mark-with-tooltip';
-export { StaticTooltip } from './components/tooltips/static-tooltip';
-export { Toggletip } from './components/tooltips/toggletip';
-export { TooltipNew } from './components/tooltips/tooltip-new';
-export { Tooltip } from './components/tooltips/tooltip'; // Tours
+export { Toggletip } from './components/tooltips/toggletip'; // Tours
 
 export { HighlightHalo } from './components/tours/highlight-halo'; // Typography
 
@@ -274,7 +269,10 @@ export { HomeIcon } from './components/icons/home-icon';
 export { TagIcon } from './components/icons/tag-icon';
 export { FileIcon } from './components/icons/file-icon';
 export { LoudspeakerIcon } from './components/icons/loudspeaker-icon';
-export { ShieldIcon } from './components/icons/shield-icon'; // ---------
+export { ShieldIcon } from './components/icons/shield-icon';
+export { PeopleIcon } from './components/icons/people-icon';
+export { StatsIcon } from './components/icons/stats-icon';
+export { SpeechBubbleIcon } from './components/icons/speech-bubble-icon'; // ---------
 // CONSTANTS
 // ---------
 
@@ -291,6 +289,7 @@ export { createEvent } from './helpers/dom/create-event';
 export { cssPropertyDistributor } from './helpers/dom/css-property-distributor';
 export { default as domElementHelper } from './helpers/dom/element-helper';
 export { nativeInputValueSetter } from './helpers/dom/native-input-value-setter';
+export { useFlexGapCheck } from './helpers/dom/use-flex-gap-check';
 export { getReactElementsByType, getReactElementsWithoutType } from './helpers/react/react-elements';
 export { default as useContinuousIntersectionObserver } from './helpers/utils/continuous-intersection-hook';
 export { default as useIntersectionObserver } from './helpers/utils/intersection-hook';